requirements.txt
# Fichier pour gérer les dépendances

# Dépendances principales
pillow == 10.3.0
customtkinter == 5.2.2
folium == 0.16.0
branca == 0.7.1
requests == 2.31.0
matplotlib == 3.8.3
pyinstaller == 6.5.0
PyQt5 == 5.15.10

# Autre dépendances
altgraph==0.17.4
build==1.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.51.0
idna==3.6
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
pefile==2023.2.7
pip-tools==7.4.1
pyinstaller-hooks-contrib==2024.3
pyparsing==3.1.2
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.2
setuptools==69.2.0
six==1.16.0
urllib3==2.2.1
wheel==0.43.0
xyzservices==2024.4.0